Largely re-implement the Budgie menu applet
This adds a new backend structure for tracking installed applications inspired by the UbuntuBudgie application-menu repo. This allows much better and easier tracking of when applications are added and removed. Updates are quite fast on my system, with no noticable lag.

The backend is taken advantage of by a re-worked UI for the menu. Views are abstracted, making it far easier to implement different types of views down the road.

A major change with the rework is that there is a new option for the menu to show or hide Control Center panels from the view. Even if they are hidden from the view, they are still searchable. By default, these items are not shown to match the current behavior of the menu.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>