Integrate functions from user-indicator applet to menu
This adds a footer to the Budgie Menu with two buttons: a user button and a power button. The user button doesn't currently do anything; its functionality will be added at a later time. The power button opens a menu with the various session and power controls.

The power menu is implemented as an overlay widget over the view widgets. This allows having a menu that doesn't resize the entire popover when it opens. However, this means that in a lot of themes, the menu will have a transparent background. The built-in theme has been modified to address this, but other themes will need editing to take this into account.

Speaking of theming, there are several new CSS classes used by various parts of the Budgie Menu:
- `budgie-menu-header`: Applied to the GtkBox holding the search entry widget
- `budgie-menu-footer`: Applied to the GtkBox holding the user and power buttons
- `budgie-menu-overlay`: Applied to the GtkListBox holding menu items. These widgets (in the event that there are more than one) are a part of a GtkOverlay as overlay widgets
- `user-icon-button`: Applied to the user button in the footer

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>