Add support for GPU switching when launching applications
This is done via switcheroo-control. If switcheroo-control is not installed or the system only has one GPU, it just starts the application as before. If those conditions are met, however, and the application asks to use either the default or non-default GPU, then the environment is applied to the launch context as appropriate.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>