Move search checking to a new RelevancyService
This creates a new centralized place to keep track of every application's relevancy relative to a search term.

It works pretty similarly to the old searching, but adds some "fuzzyness" by calculating the Levenshtein Distance if the term doesn't quite match an app's name. Apps in the menu are filtered and sorted based on this calculated relevancy, with filtering using an arbitrary threshold.

It should be easier to tweak how relevancy is calculated going forward with this new API.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>